% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_wtg.R
\docType{data}
\name{geom_wtg}
\alias{geom_wtg}
\alias{GeomWtg}
\title{World Tile Grid Geom}
\format{An object of class \code{GeomWtg} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
geom_wtg(mapping = NULL, data = NULL, border_col = "white",
  border_size = 0.125, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

GeomWtg
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or
\code{aes_()}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{ggplot()}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{fortify()} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{border_col}{border color of the state squares, default "\code{white}"}

\item{border_size}{thickness of the square state borders}

\item{...}{other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{borders()}.}
}
\description{
Pass in a data frame of countries (iso2c, i23c, name) and a value column and
get back a world tile grid.
}
\details{
\strong{IMPORTANT} : For now, you need to pass in a \emph{complete} set of countries
(the values can be \code{NA}). When I get time I'll work on this limitation but
there's a \link{wtg} data frame exported from the package that you can use
to merge with your data to ensure you've got all the tiles.

\strong{ALSO} : Labeling world tile grids is a tricky business and no labeling
parameters are planned for this since you should think very carefully about
the tradeoffs of tiny text/numbers vs readability. These charts are really
only good for overviews in single-chart form or highlighting stark differences
in panel-form.

\cr
There are two special/critical \code{aes()} mappings:\cr
\cr
\itemize{
\item \code{country} (so the geom knows which column to map the country names/abbrevs to)
\item \code{fill} (which column you're mapping the filling for the squares with)
}
}
\keyword{datasets}
