% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtilegrid-package.R
\docType{package}
\name{worldtilegrid}
\alias{worldtilegrid}
\alias{worldtilegrid-package}
\title{A ggplot2 Geom for World Tile Grids}
\description{
\itemize{
\item Ref: \url{https://policyviz.com/2017/10/12/the-world-tile-grid-map/}
\item Ref: \url{http://www.maartenlambrechts.com/2017/10/22/tutorial-a-worldtilegrid-with-ggplot2.html}
\item URL: \url{https://gitlab.com/hrbrmstr/worldtilegrid}
\item BugReports: \url{https://gitlab.com/hrbrmstr/worldtilegrid/issues}
}
}
\author{
Bob Rudis (bob@rud.is)

Maarten Lambrechts

Jon Schwabish
}
