% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-wtg.R
\name{theme_enhance_wtg}
\alias{theme_enhance_wtg}
\title{World tile grid theme cruft remover that can be used with any other theme}
\usage{
theme_enhance_wtg()
}
\description{
Removes:
}
\details{
\itemize{
\item panel grid
\item all axis text
\item all axis ticks
\item all axis titles
}
}
