% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtilegrid-package.R
\docType{package}
\name{worldtilegrid}
\alias{worldtilegrid}
\alias{worldtilegrid-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/worldtilegrid}
\item BugReports: \url{https://gitlab.com/hrbrmstr/worldtilegrid/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
