% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldtilegrid-package.R
\docType{package}
\name{worldtilegrid}
\alias{worldtilegrid}
\alias{worldtilegrid-package}
\title{A ggplot2 Geom for World Tile Grids}
\description{
A "tile grid map" is a cartogram that uses same-sized tiles
in approximate, relative positions of each other to represent a world
map. The world tile grid relative position reference system used by
this 'ggplot2' 'Geom/Stat' was the original work of 'Jon Schwabish' and
converted to 'CSV' by 'Maarten Lambrechts'.
}
\details{
\itemize{
\item Ref: \url{https://policyviz.com/2017/10/12/the-world-tile-grid-map/}
\item Ref: \url{http://www.maartenlambrechts.com/2017/10/22/tutorial-a-worldtilegrid-with-ggplot2.html}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/worldtilegrid}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/worldtilegrid/issues}
}

}
\author{
Bob Rudis (bob@rud.is)

Maarten Lambrechts

Jon Schwabish
}
